A Simple Android Project Demonstrating Persistence of Inter-Related Objects Using JDXA ORM
This project shows a simple Android app using JDXA ORM to exchange data
of the inter-related objects with an on-device SQLite relational
database. The object model used by this app has one-to-one and
one-to-many relationships. A company has many departments and many
foreign locations. An employee has an address and works in a department.